Introduce separate bulkdata config params for file export
Previously, the file-based export would use the COS properties for
checkpointing. This is a bit confusing and, additionally, I think we'd
want a different default for "when to start writing" (in COS there are
constraints on the size and count of the parts, while writing to local
disk has different concerns).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>